# Writer-Assistant/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WriteAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Document {
  id: number;
  title: string;
  content: string;
  userId: string;
  wordCount: number;
  /** @nullable */
  goalWordCount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface DocumentInput {
  /** @minLength 1 */
  title: string;
  content?: string;
}

export interface DocumentUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  /** @nullable */
  goalWordCount?: number | null;
}

export interface DocumentStats {
  totalDocuments: number;
  totalWords: number;
  recentDocuments: Document[];
}

export interface DocumentVersion {
  id: number;
  documentId: number;
  title: string;
  content: string;
  wordCount: number;
  /** @nullable */
  label?: string | null;
  createdAt: string;
}

export interface DocumentVersionInput {
  title: string;
  content: string;
  wordCount: number;
  /** @nullable */
  label?: string | null;
}

export type WorldEntityType = typeof WorldEntityType[keyof typeof WorldEntityType];


export const WorldEntityType = {
  character: 'character',
  place: 'place',
  item: 'item',
} as const;

export type WorldEntityFields = { [key: string]: unknown };

export interface WorldEntity {
  id: number;
  documentId: number;
  userId: string;
  type: WorldEntityType;
  name: string;
  fields: WorldEntityFields;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type WorldEntityInputType = typeof WorldEntityInputType[keyof typeof WorldEntityInputType];


export const WorldEntityInputType = {
  character: 'character',
  place: 'place',
  item: 'item',
} as const;

export type WorldEntityInputFields = { [key: string]: unknown };

export interface WorldEntityInput {
  type: WorldEntityInputType;
  name: string;
  fields?: WorldEntityInputFields;
  /** @nullable */
  imageUrl?: string | null;
}

export type WorldEntityUpdateFields = { [key: string]: unknown };

export interface WorldEntityUpdate {
  name?: string;
  fields?: WorldEntityUpdateFields;
  /** @nullable */
  imageUrl?: string | null;
}

export type AiSuggestInputType = typeof AiSuggestInputType[keyof typeof AiSuggestInputType];


export const AiSuggestInputType = {
  improve: 'improve',
  expand: 'expand',
  shorten: 'shorten',
  rephrase: 'rephrase',
  continue: 'continue',
} as const;

export interface AiSuggestInput {
  text: string;
  type: AiSuggestInputType;
  /** @nullable */
  context?: string | null;
}

export interface AiSuggestResult {
  suggestion: string;
}

export interface AiGrammarInput {
  text: string;
}

export type GrammarErrorType = typeof GrammarErrorType[keyof typeof GrammarErrorType];


export const GrammarErrorType = {
  grammar: 'grammar',
  spelling: 'spelling',
  style: 'style',
} as const;

export interface GrammarError {
  message: string;
  offset: number;
  length: number;
  type: GrammarErrorType;
  /** @nullable */
  suggestion?: string | null;
}

export interface AiGrammarResult {
  errors: GrammarError[];
  correctedText: string;
}

export type AiImageInputSize = typeof AiImageInputSize[keyof typeof AiImageInputSize];


export const AiImageInputSize = {
  '1024x1024': '1024x1024',
  '1536x1024': '1536x1024',
  '1024x1536': '1024x1536',
} as const;

export interface AiImageInput {
  prompt: string;
  size?: AiImageInputSize;
}

export interface AiImageResult {
  b64_json: string;
}

export interface AiSummarizeInput {
  text: string;
  /** @nullable */
  title?: string | null;
}

export interface AiSummarizeResult {
  summary: string;
}

export interface AiPrologueInput {
  text: string;
  /** @nullable */
  title?: string | null;
}

export interface AiPrologueResult {
  prologue: string;
}

export type ListWorldEntitiesParams = {
type?: ListWorldEntitiesType;
};

export type ListWorldEntitiesType = typeof ListWorldEntitiesType[keyof typeof ListWorldEntitiesType];


export const ListWorldEntitiesType = {
  character: 'character',
  place: 'place',
  item: 'item',
} as const;