# Writer-Assistant/lib/api-zod/src/generated/types/aiGrammarResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WriteAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GrammarError } from './grammarError';

export interface AiGrammarResult {
  errors: GrammarError[];
  correctedText: string;
}