# Writer-Assistant/lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WriteAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  title: string;
  content: string;
  userId: string;
  wordCount: number;
  /** @nullable */
  goalWordCount?: number | null;
  createdAt: string;
  updatedAt: string;
}