# Writer-Assistant/lib/api-zod/src/generated/types/documentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WriteAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  /** @nullable */
  goalWordCount?: number | null;
}