# Writer-Assistant/lib/api-zod/src/generated/types/documentVersionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WriteAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentVersionInput {
  title: string;
  content: string;
  wordCount: number;
  /** @nullable */
  label?: string | null;
}