# Writer-Assistant/lib/api-zod/src/generated/types/worldEntity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WriteAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorldEntityFields } from './worldEntityFields';
import type { WorldEntityType } from './worldEntityType';

export interface WorldEntity {
  id: number;
  documentId: number;
  userId: string;
  type: WorldEntityType;
  name: string;
  fields: WorldEntityFields;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}